+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for ?limit
 - `findModel` supports all types rather than just int now.
 - Default `getRouteKey` implementation
+- Simple file logging implementation
 
 ### Changed
 - OPTIONS with ID requested via query string now when using `Builder`